+++ src/Components/common/Header/BasicHeader.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import prev from '../../Assets/icons/icon-arrow-back.svg';
-import more from '../../Assets/icons/icon-more-vertical.svg';
+import prev from '../../../Assets/icons/icon-arrow-back.svg';
+import more from '../../../Assets/icons/icon-more-vertical.svg';
 import styled from 'styled-components';
-import HeaderLayout from '../../Styles/HeaderLayout';
+import HeaderLayout from '../../../Styles/HeaderLayout';
 import { useNavigate } from 'react-router-dom';
 
 const BasicHeader = (props) => {
@@ -15,7 +15,7 @@ const BasicHeader = (props) => {
             navigate(-1);
           }}
         />
-        <div>{props.children}</div>
+        {props.children && <div>{props.children}</div>}
       </ContentLayout>
       <MoreButton />
     </HeaderLayout>
