+++ src/Components/common/Header/MainHeader.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import HeaderLayout from '../../Styles/HeaderLayout';
+import HeaderLayout from '../../../Styles/HeaderLayout';
 import logo from '../../Assets/logo-Alphabet.svg';
 import search from '../../Assets/icons/icon-search.svg';
 import styled from 'styled-components';
